+++ packages/dataviews-client/src/resources/index.ts
@@ -1,0 +1,2 @@
+export * from './resources-slice'
+export { default as resolveResourceEndpoint } from './resources-endpoint'